refactor Runner#serial and Runner#concurrent to accept tests in arguments
Before this change, code would be duplicated a lot. For example,
before/after hooks require the same test run functionality as implemented
in Runner#serial. But because previously Runner#serial was hard-coded to
get tests from `this.tests.serial` property, it was not possible to use it
to run before/after hooks.

With this change, one can `this.serial([testA, testB])` to run tests.