Fixup coverage.
Due to NYC defaults, we were missing coverage on lib/test.js

I also added `npm run coverage` as a build target, which is for use
on developer machines. It launches a browser window with coverage
from the last run.

We no longer need NYC as a dev dependency. `tap` handles it for us.

`maintaining.md` updated to reflect all these changes.