Integrate time-require when `debug` is enabled.
I want to start focusing on getting our require times down,
both in the main thread and forked tests.

This adds `time-require` support directly to the CLI to assist in that.
This is especially important for forked test processes, where I currently
have to manually add a line of code each time I want to test an optimization.

I don't think this should be a publicly documented move. I think we want
the freedom to remove this once we are happy with performance.

I have placed comments above all `time-require` references, indicating
it is for internal use only. Users have been given fair warning against
relying on its inclusion long term.

Usage:

```sh
$ DEBUG=ava ./cli.js test/fixture/es2015.js --sorted
```

`--sorted` is optional. It sorts the requires from most expensive to least. The default is sequential sorting.

The output will display multiple `time-require` reports.
One for each forked process, and the last one for the main thread.