improve source-map tests
Use the source-map-fixtures package to provide a fixture which throws an error.
Stack traces should be corrected for this fixtures.

Add a script to generate a test fixture that comes with an input source map.
Errors from this fixture should be mapped to the input source. Having the script
makes it easier to update this fixture in the future.

The generated fixture now uses a source map file. Commenting out the
inputSourceMap code in lib/babel.js didn't cause the test to fail if an inline
source map was used. I'm not sure why that is. The test fails as expected when
using a map file.

Note that the original 'stack traces for exceptions are corrected using a source
map' test only tested the source map cached by lib/babel.js. This is already
covered by the new 'stack traces for exceptions are corrected using a source
map, taking an initial source map for the test file into account' test.

We throw the error from another fixture which uses a source map file. This
implicitly tests source-map-support. We're not testing a fixture which uses an
inline source map, firstly because that's somewhat superfluous and secondly
because the generated stack trace line is deemed not important due to the
logger's behavior and a source-map-support bug
(<https://github.com/evanw/node-source-map-support/pull/119>).