(re)run specific tests when watching
The watcher checks whether the changed files affect just tests. If so, it runs
those tests again. Else it reruns the entire test suite.

The API now resolves the default files patterns when it's constructed. It also
exposes the default exclusion patterns. This allows the watcher to check if a
changed file was indeed a test.