consistent error determination
The child processes determine whether the test had an error based on its `error`
property not being `undefined`. However they then change this property to an
empty object if there was no error.

The API determines whether a test had an error based on its (empty) error object
having a `message` property. If a test did not have an error, its `error`
property is set to `null`. This prevents errors without messages from being
reported correctly.

Instead set `error` to `null` in the child processes and rely on that in the
API.

I've added a test to validate errors without messages get reported.