ensure test errors are errors
If a test throws synchronously, but the error is not an Error instance, create
an AssertionError like for rejected promises.

Use inspect() to include a string representation of the actual error, and
include it in the `actual` property.

Remove special handling which converted `undefined` error values to the
`'undefined'` string.