don't pause stderr in watcher test
In Node 0.10 pause() does not return the stream.

The stream was being paused because get-stream did not use flowing mode, and
adding the 'data' listener would switch the stream to flowing mode. However
that did not work with Node 0.10 either. get-stream has been updated to use
flowing mode, which makes pause() unnecessary. The test now passes in Node 0.10.