handle exceptions when initially running files
Catch exceptions when initially running files. Don't run any tests, just report
the exception.

Move test teardown into a callback for the run promise so it can be used when
--match is used but there are no matches, and when exceptions occur while
initially running files.

Teardown causes the test promise to reject, which leads to an attempt to run
the tests. Add a guard to prevent this.

Fixes #622.