Cleaned up experimental concurrency pool, renamed CLI flag.
Renamed the CLI flag to `--concurrency, -c` as per @sindresorhus suggestion. Marked it as experimental instead of beta.

Updated the new concurrency pool runner to force pool size to 1 when `--serial` flag is set as per @novemberborn suggestion.

Reduced the nesting complexity of the concurrency pool code and broke up portions of the logic to ensure readability.