Improve the issue template.
The current issue template uses HTML comment tags to remove template text from the rendered output. I have gone through the last dozen issues filed, and it appears everyone deletes that anyways. By removing the comment tags, people can click "Preview" and get a pretty print of the template markdown.

Reformat lists and links. Tries to get links as close to the left edge as possible, so they are easy to see, and copy-paste from the raw markdown editor. Puts multiple links on their own line for the same reason.

Adds a request that they explicitly tell us what their AVA and `npm` versions are.

Adds a section for their AVA config and CLI args.

Adds a section for linking to thier project if public.

Requests they create a minimal reproduction and post it to GitHub.

Closes #787