Update JSPM+SystemJS recipe per @jamestalmage.
See istanbuljs/nyc#305 - specifically [this comment and the one after it](https://github.com/istanbuljs/nyc/issues/305#issuecomment-231517122).

Essentially, @jamestalmage brought up that it would be good to explicitly mention in the recipe that you should not pre-build your project with JSPM/SystemJS to run your tests. The purpose of ava-jspm-loader is that you do not have to build before testing.

Closes #954