Fix tests in Node 6.5.0 by handling `fn` test name
Node 6.5.0 uses a newer version of the v8 engine that is a bit smarter
in figuring out the name of a function. Therefore, some tests were
breaking because we presumed `test.name` to be a falsy value, and now
it's `'fn'`. This PR handles `'fn'` the same way it handles a falsy function
name value. While it fixes the breaking tests, it may not be the best
choice in the long term.