Fix: Flow type ambiguity
Fixes #1114

As reported in #1114, flow reports ambiguity when defining tests:

```js
test('my test', t => {})
```

Flow will report that it could apply either the `Test` type or the `Macro` type. The `Test` type is an obvious fit, but `Macro` also applies since the `...args` resolves to the empty array, and the `title?` attribute happens to be not provided.

My suggested fix is to merge the `Test` and `Macro` types together since `Macro` is really a superset of the `Test` type, and we don't lose any type safety by just expecting to always get Macros. This super set type is called `TestImplementation` which better matches how AVA names these things.

I tested this locally on a flow-typed repository and confirmed that the type ambiguity errors are solved.