Add flow type tests (#1230)
Closes #1224.

This adds flow-bin, adds it to the test running script, and adds some regression tests based on prior issues.

In an immediate proof of value (and proof of my own prior sloppiness) it spotted an issue that #1219 sought to fix and did not do so. Fix included along with test.