Format values when AssertionErrors are created
Format statements, generate diffs and otherwise format values when
AssertionErrors are created. Make labels contextual to the assertion.

Always print power-assert statements in mini and verbose reporters, but
after the values. Accept the duplication when the statement covers the
value passed to the assertion.

Detect whether `t.throws()` failed because no error was thrown, or because
the thrown error did not meet expectations.

Explicitly test `t.fail()` default message in assertion test.