Refactor Test implementation
* Do away with underscore prefixes. They were used inconsistently, and
are generally not worth it

* Keep `_test` prefixed in ExecutionContext, since that is exposed to
calling code

* Reorder methods

* Assume all arguments are passed and are correct. They're already
validated in `test-collection.js`

*Pass metadata when instantiating Tests

* Rewrite test finish logic. There's still a fair bit of interleaving
due to the `test.cb()` API, but it's now a lot easier to understand.

Due to the last change, tests with `t.end()` and only synchronous
assertions end immediately. Previously they would end asynchronously
due to a promise being in the completion chain.

Similarly, returning a promise or observable for a `test.cb()` test
immediately fails.