Track snapshot files touched during test run, ignore in watcher
Workers can emit which files were touched during the test run. This is
used to communicate to the watcher which snapshot file events to ignore
from the current test run, stopping the watcher from running the same
tests repeatedly.

Note that this is only an issue when the `sources` glob has been
customized by the user. The default glob excludes snapshot files.

Files are ignored only once, so subsequent edits of snapshot files
(e.g. by reverting a commit) will cause tests to be rerun.