Use --verbose when testing CLI output
It cannot be guaranteed that all lines output by the mini reporter are
captured in the stderr of the CLI invocation during testing. Use
--verbose when matching output that the mini reporter may otherwise
rewrite.

Somewhere on at least Node.js 8.2.1, when the package-lock.json is
discarded prior to installation, something in our dependency tree is
different which results in the test title no longer being in the stderr
of the AVA CLI.

Compare:

```
⠋ ESC[2KESC[1AESC[2KESC[G
⠙ ESC[2KESC[1AESC[2KESC[G
⠹ ESC[2KESC[1AESC[2KESC[G
⠸ ESC[2KESC[1AESC[2KESC[G
⠼ ESC[2KESC[1AESC[2KESC[G
⠴ ESC[2KESC[1AESC[2KESC[G
⠦ ESC[2KESC[1AESC[2KESC[G
⠦ test › works…

 1 passedESC[2KESC[1AESC[2KESC[1AESC[2KESC[1AESC[2KESC[G
 1 passed [18:32:54]
```

Versus:

```
⠋ ESC[2KESC[1AESC[2KESC[G
⠙ ESC[2KESC[1AESC[2KESC[G
⠹ ESC[2KESC[1AESC[2KESC[G
⠸ ESC[2KESC[1AESC[2KESC[G
⠼ ESC[2KESC[1AESC[2KESC[G
⠴ ESC[2KESC[1AESC[2KESC[G
⠴ …

 1 passedESC[2KESC[1AESC[2KESC[1AESC[2KESC[1AESC[2KESC[G
⠦ …

 1 passedESC[2KESC[1AESC[2KESC[1AESC[2KESC[1AESC[2KESC[G
 1 passed [18:35:49]
```