Edit copy
* Consistent spelling of Express, Mongoose and SuperTest
* Move prerequisites into their own section, so the test file section
  reads better
* Reformat code examples
* Remove code comments from examples where they repeat the preceding
paragraph
* Reword paragraphs so they read better (in my humble opinion, that is)

I've also changed the section on using `test.serial`: I'm pretty sure
Mongoose already uses the same connection in the tests and the app.
Notably, the connection is made in a `test.before()` hook, not a
`test.beforeEach()` hook. Thus the real concern is tests interleaving
and the database being in an unknown or conflicting state.