Assume npm@5.4, work around permissions issue (#1499)
* Use npm@5.4.0 in CI

* Use --no-shrinkwrap option rather than removing package-lock.json

* Rebuild package-lock.json with npm@5.4.0

* Make flow binary executable when running tests

Since npm@5.4.0, installing `flow-bin` results in non-executable
binaries being installed. As a workaround, mark the binary as executable
when our non-Windows test script is run.

* Prefer offline mode when installing locked deps in AppVeyor

* Work around npm/Windows issues in AppVeyor builds

Skip optional dependencies, which is a workaround for
https://github.com/npm/npm/issues/17671. Still, that workaround doesn't
seem to work when package-lock.json files are used. Thus only test fresh
dependencies on AppVeyor for now.