Simplify reported value
When the value of an uncaught exception is a function, AVA previously
displayed the source of that function in the error report. During the
review process for an unrelated change, this condition was found to be
too rare to justify the maintenance overhead. Do not attempt to display
the source of function values in these cases.