Change Babel options used for test compilation
Always extend the project's Babel configuration, if any.

Always apply the `@ava/stage-4` preset. Fixes #1488. Users can disable
this preset by adding it to the AVA's Babel configuration with the
`false` option:

```
"ava": {
	"babel": {
		"presets": [
			["ava/stage-4", false]
		]
	}
}

And yes, `ava/stage-4` so now an alias for the preset, which means it
can be used even if the preset is not installed as a top-level
dependency. Fixes #1225.

`@ava/transform-test-files` is applied as part of the test compilation,
but cannot be overridden by users like `@ava/stage-4` can.