Refactor TypeScript definition
Remove the need for a build step.

Update the `throws` and `notThrows` assertions, adding `Observable` and
allowing the returned error to be typed.

Remove the inline documentation, to be added later.

Simplify typing of `t.context`. Reassigning the `test` method along with
a type cast is now sufficient.

Update to TypeScript 2.7.1.