Don't reload Babel config in watch mode
Babel config is hard to reload. Plugins and presets may be in the
require cache. AVA may detect changes and update its cache keys, but
because of the require cache the compilation output may not take the
changes into account. This ends up polluting the compilation cache.

Change the API implementation so it caches the Babel config. This only
impacts watch mode.