Provide getColorDepth() shim for fake TTYs
The AssertionError constructor in Node 10 depends on the TTY
interface. More specifically, if isTTY is true on stdout then
getColorDepth() must also be available. Ava provides a shim for
isTTY in the test workers but does not provide a getColorDepth()
implementation. This leads to unexpected failures when code under
test is using Node's built-in assert module.