Use source-mapped file to build snapshots
* Use the filename of the source file to build the snapshot filename
* Put the relative path of the actual source file in the Markdown snapshot

This is great for TypeScript projects where AVA is applied to pre-build files. The snapshots will be named `file.ts.snap`, and the relative path will be correct, rather than pointing at the output directory.

This is a breaking change, since AVA will now write new snapshot files, without removing the previous ones. It may not catch bugs introduced at the same time as the project is updated to the new AVA version.