Refactor assertions to be a class
This changes assertions to be properties on the class. Before,
assertions were functions which needed to be bound to the test instance.
This led to the fact that snapshot() assertion accessed property which
only existed on Test - to compare snapshots. Now, the functions which
assertions access are clearly defined as those that need to be
implemented by child class.

This change also updates how ExecutionContext is created, as well as
minor changes to the test/assert.js.