Remove built-in Babel support
This updates code and tests to remove built-in Babel support. In the
future, it needs to be opted in to, and @ava/babel needs to be installed
besides AVA itself.

Our tests have become quite muddled, and some of them require Babel to
still be enabled, but that's all OK for now.