Refactor globbing
This removes the concept of "helpers" from AVA. Instead, the Babel
provider can be configured to compile additional files as tests. Files
starting with an underscore are still *not* treated as tests.

Now that AVA itself will no longer be compiling *anything*, we've also
removed the concept of "sources". Instead you can specify glob patterns
for files that the watcher should ignore.

The concept of helpers persists in our ESLint plugin however.

Glob patterns are always relative to the project directory.