Use CJS syntax in documentation
Out of the box, ESM syntax will only work in Node.js 13, so having that
as the default example is misleading.

Leave it for the TypeScript recipe since TypeScript doesn't work out of
the box anyway. Leave the config example, which will be updated when we
update the config formats for ESM.