Retain (most of) Node.js internals in stack traces
This retains most of the Node.js internals in stack traces for the mini and verbose reporters. Internals are dimmed in the output. Call sites are prefixed with a small pointer.

Reporter logs are now tested for each Node.js version separately.

Stack beautification now takes place in the main process, not the workers.

Fixes #2110.

Co-authored-by: Mark Wubben <mark@novemberborn.net>