Force not-ci in snapshot-order tests
When ava is run in CI, new snapshots cannot be written. In normal
circumstances, this ensures that snapshot assertions don't get merged
without a recorded snapshot; lacking this feature, CI could appear to
pass when in fact snapshot assertions were not checked. In this case,
however, writing snapshots is intended.