Test edge case involving new t.snapshot.skip()s
When new t.snapshot.skip() assertions are followed by a t.snapshot()
assertion, the run will fail with a RangeError, as a result of
attempting to save nonconsecutive snapshots. This can only happen when
the t.snapshot.skip()s are wholly new, not when they are the result of
skipping an existing snapshot assertion.