Record blanks when skipping snaps without data
This occurs when the old snapshots are unreadable or `t.snapshot.skip()`
is added at the end of a test.

Previously, skipping snapshots when the old snapshots are unreadable
would throw either SnapshotError or 'Can't skip snapshot assertions when
updating snapshots'. Adding `t.snapshot.skip()` to the end of a test
would throw RangeError.

Blanks are formatted as `<No Data>` in snapshot reports.

Blanks compare equal to anything and are filled in when a value is
passed; essentially, they behave like they didn't exist.

Internally, blanks are represented with `{buffer: undefined}`.