Move snapshot data reformatting into load, save
The blocksByTitle format is an implementation detail of Manager, whereas
the CBOR data structure is a characteristic of the snapshot file
protocol.

This change reduces codepath duplication between generateReport() and
encodeSnapshots(), and avoids passing blockIndices around everywhere.

If decodeSnapshots() and encodeSnapshots() are later exposed as a
library, this may be closer to the desired interface.