Remove trim-off-newlines
Concordance-formatted text shouldn't have newlines at the start or end.
I'm not sure about power-assert output, but let's assume the same.

An extra newline around TypeScript diagnostic text is fine — this is only a
convenience for in-process TypeScript compilation which requires a lot of custom
configuration to be a thing.

For the remaining use case, trim() should suffice.