Find ava.config.* files outside of project directory
Fixes #2285.

Find configuration files in parent directories, until a `.git` is
encountered.

Print a warning when AVA is run with a configuration file that is not
next to the `package.json`. This is now allowed when passing `--config`.

Update the documentation to reflect these changes. Remove documentation
of the experimental "next-gen" feature in AVA 3.