Update readme.md
Long time user, first time comitter!

This badge is a part of @techforpalestine's mission to help humanize the people of Palestine.
Adding a badge is way to show support within the tech community for peace in Gaza, and for an end to violence against civilians.

I was inspired to make this change by the call to action to support Vadim and the people of Ukraine. As of Jan 2, 2024, [more than 22,000 Palestinians have been killed, with almost 60,000 injured](https://www.aljazeera.com/news/2024/1/2/israeli-bombardment-ongoing-as-death-toll-surpasses-22000). 

For anyone unfamiliar with Palestine, I recommend Paul Biggar's blog post titled [I can't sleep](blog.paulbiggar.com/i-cant-sleep), which is a well referenced post about the history of Palestinian oppression and the current day situation. That blog post also resulted in the forming of @techforpalestine. 

The badge used in this commit will not change the message. It will only say "Ceasefire Now". However, we could choose to instead use `[![Ceasefire Now](https://badge.techforpalestine.org/default)](https://techforpalestine.org/learn-more)
`, which will change if a ceasefire is achieved, to indicate the next goal for Palestinian freedom.  @techforpalestine promises that:

> The message will always be carefully calibrated so that you don't need to be an expert in politics or foreign policy to support, and will always be positive and support the humanity of the Palestinian and Israeli people.