Bump `@vercal/nft` from 0.27.5 to 0.29.2
Upgrade AVA's dependency on `@vercal/nft` from 0.27.5 to the latest
version, 0.29.2. This transitively resolves a dependency on a deprecated
dependency, thus improving the dependency tree of AVA. In particular,
the transitive dependency on the `glob` package is bumped from 7.2.3 to
10.4.5.

This change can be verified by running `npm clean-install --omit dev` on
the main branch and this change and comparing the output. As of writing,
the former outputs multiple deprecation warnings whereas the latter has
zero deprecation warnings (of course, this is a time sensitive operation
so your results may vary).