Remove special .only() behavior in watch mode
Remove special .only() behavior in watch mode.

This is a historical feature back from when AVA tried to persist "runExclusive" during regular runs. This behavior
conflicts with interactive mode, and in any case interactive mode makes it easier to focus on a specific test file or title.

Implement --match using the selected flag. This is a newer flag, previously used for line number selection. It's a better choice than overloading the `exclusive` flag used with `.only()`.

Together these changes simplify the runner logic, allowing for the change of behavior.