Release process tweaks
* Verify release commit is in the main branch

Typically the tag is pushed first, CI runs, clearing the way for 'main' to be pushed. However we must not release until this is done.

* Improve CI status check

Execute logic within jq; look for any passing workflow and fail if there are none.

* Execute package.json check within jq

No need for Bash, jq can handle this.

* Perform package.json check first

It's the most important check, are we publishing what we think we're publishing.

* Update instructions